Add pm, a dual-purpose tool:  - Persistent IRC server monitor  - DNS pool management tool (hence the name 'pm')
The idea for this project was to monitor a network of IRC servers actively,
by maintaining persistent clients, and then attempt to heuristically
(and naively) determine if a server is 'down.' If a server is down, we
want to extricate it from the pool.

There is a separate frontend to just do pool management, so humans can
manage the DNS pool without having to futz with zone files and rndc.

As a pool manager, this project was successful. As a server monitor, it was
frought with heuristic problems. This needed to be rethought, but I ended
my involvement with EsperNet before a chance to rethink came about.